+++ src/components/Fav/FavHeader/FavHeader.tsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
 import Typography from '@mui/material/Typography';
 
-import { UseFavP } from '../hooks/useFavPaginated';
+import { UseFavP } from '../hooks/usePlaylistPaginated';
 import { skinPreset } from '../../../styles/skin';
 import RandomGIFIcon from './RandomGIF';
 import FavSettingsButtons from '../FavSetting/FavSettingsButton';
